Addresses #946, adds a temporary fix that prevents crashing on *my machine*. This should be explored and tested ASAP.
git-svn-id: http://developer.mangband.org/svn/mangband/trunk@1364 802d84f3-b497-254e-bede-06c0a5a3456a

+++ src/client/netclient.c
@@ -415,7 +415,7 @@ int Net_packet(void)
 	/* Process all of the received client updates */
 	while (rbuf.buf + rbuf.len > rbuf.ptr)
 	{
-		type = (*rbuf.ptr & 0xFF);
+		type = (byte)(*rbuf.ptr & 0xFF);
 		if (ack_tbl[type] == NULL)
 		{
 			errno = 0;
